! consolidate all manage signature functions to common sub ! add a enforce number of smiles to admin signature maintenance
