+++ sources/subs/ManageFeatures.subs.php
@@ -15,340 +15,6 @@
  *
  */
 
-use ElkArte\Helper\Util;
-
-/**
- * Loads the signature from 50 members per request
- * Used in ManageFeatures to apply signature settings to all members
- *
- * @param int $start_member
- * @return array
- */
-function getSignatureFromMembers($start_member)
-{
-	$db = database();
-
-	$members = [];
-
-	$db->fetchQuery('
-		SELECT 
-			id_member, signature
-		FROM {db_prefix}members
-		WHERE id_member BETWEEN ' . $start_member . ' AND ' . $start_member . ' + 49
-			AND id_group != {int:admin_group}
-			AND FIND_IN_SET({int:admin_group}, additional_groups) = 0',
-		[
-			'admin_group' => 11,
-		]
-	)->fetch_callback(
-		function ($row) use (&$members) {
-			$members[$row['id_member']]['id_member'] = $row['id_member'];
-			$members[$row['id_member']]['signature'] = $row['signature'];
-		}
-	);
-
-	return $members;
-}
-
-/**
- * Updates the signature from a given member
- *
- * @param int $id_member
- * @param string $signature
- */
-function updateSignature($id_member, $signature)
-{
-	require_once(SUBSDIR . '/Members.subs.php');
-	updateMemberData($id_member, ['signature' => $signature]);
-}
-
-/**
- * Update all signatures given a new set of constraints
- *
- * @param int $applied_sigs
- */
-function updateAllSignatures($applied_sigs)
-{
-	global $context, $modSettings;
-
-	require_once(SUBSDIR . '/Members.subs.php');
-
-	// This is horrid - but I suppose some people will want the option to do it.
-	$done = false;
-	$context['max_member'] = maxMemberID();
-
-	// Load all the signature settings.
-	list ($sig_limits, $sig_bbc) = explode(':', $modSettings['signature_settings']);
-	$sig_limits = explode(',', $sig_limits);
-	$disabledTags = !empty($sig_bbc) ? explode(',', $sig_bbc) : [];
-
-	// @todo temporary since it does not work, and seriously why would you do this?
-	$disabledTags[] = 'footnote';
-
-	while (!$done)
-	{
-		// No changed signatures yet
-		$changes = [];
-
-		// Get a group of member signatures, 50 at a clip
-		$update_sigs = getSignatureFromMembers($applied_sigs);
-
-		if (empty($update_sigs))
-		{
-			$done = true;
-		}
-
-		foreach ($update_sigs as $row)
-		{
-			// Apply all the rules we can realistically do.
-			$sig = strtr($row['signature'], ['<br />' => "\n"]);
-
-			// Max characters...
-			if (!empty($sig_limits[1]))
-			{
-				$sig = Util::substr($sig, 0, (int) $sig_limits[1]);
-			}
-
-			// Max lines...
-			if (!empty($sig_limits[2]))
-			{
-				$count = 0;
-				$str_len = strlen($sig);
-				for ($i = 0; $i < $str_len; $i++)
-				{
-					if ($sig[$i] === "\n")
-					{
-						$count++;
-						if ($count >= $sig_limits[2])
-						{
-							$sig = substr($sig, 0, $i) . strtr(substr($sig, $i), ["\n" => ' ']);
-						}
-					}
-				}
-			}
-
-			// Max text size
-			if (!empty($sig_limits[7]) && preg_match_all('~\[size=([\d\.]+)?(px|pt|em|x-large|larger)?~i', $sig, $matches) !== false && isset($matches[2]))
-			{
-				// Same as parse_bbc
-				$sizes = [1 => 0.7, 2 => 1.0, 3 => 1.35, 4 => 1.45, 5 => 2.0, 6 => 2.65, 7 => 3.95];
-
-				foreach ($matches[1] as $ind => $size)
-				{
-					$limit_broke = 0;
-
-					// Just specifying as [size=x]?
-					if (empty($matches[2][$ind]))
-					{
-						$matches[2][$ind] = 'em';
-						$size = $sizes[(int) $size] ?? 0;
-					}
-
-					// Attempt to allow all sizes of abuse, so to speak.
-					if ($matches[2][$ind] == 'px' && $size > $sig_limits[7])
-					{
-						$limit_broke = $sig_limits[7] . 'px';
-					}
-					elseif ($matches[2][$ind] == 'pt' && $size > ($sig_limits[7] * 0.75))
-					{
-						$limit_broke = ((int) $sig_limits[7] * 0.75) . 'pt';
-					}
-					elseif ($matches[2][$ind] == 'em' && $size > ((float) $sig_limits[7] / 16))
-					{
-						$limit_broke = ((float) $sig_limits[7] / 16) . 'em';
-					}
-					elseif ($matches[2][$ind] != 'px' && $matches[2][$ind] != 'pt' && $matches[2][$ind] != 'em' && $sig_limits[7] < 18)
-					{
-						$limit_broke = 'large';
-					}
-
-					if ($limit_broke)
-					{
-						$sig = str_replace($matches[0][$ind], '[size=' . $sig_limits[7] . 'px', $sig);
-					}
-				}
-			}
-
-			// Stupid images - this is stupidly, stupidly challenging.
-			if ((!empty($sig_limits[3]) || !empty($sig_limits[5]) || !empty($sig_limits[6])))
-			{
-				$replaces = [];
-				$img_count = 0;
-
-				// Get all BBC tags...
-				preg_match_all('~\[img(\s+width=([\d]+))?(\s+height=([\d]+))?(\s+width=([\d]+))?\s*\](?:<br />)*([^<">]+?)(?:<br />)*\[/img\]~i', $sig, $matches);
-
-				// ... and all HTML ones.
-				preg_match_all('~&lt;img\s+src=(?:&quot;)?((?:http://|ftp://|https://|ftps://).+?)(?:&quot;)?(?:\s+alt=(?:&quot;)?(.*?)(?:&quot;)?)?(?:\s?/)?&gt;~i', $sig, $matches2, PREG_PATTERN_ORDER);
-
-				// And stick the HTML in the BBC.
-				if (!empty($matches2))
-				{
-					foreach ($matches2[0] as $ind => $dummy)
-					{
-						$matches[0][] = $matches2[0][$ind];
-						$matches[1][] = '';
-						$matches[2][] = '';
-						$matches[3][] = '';
-						$matches[4][] = '';
-						$matches[5][] = '';
-						$matches[6][] = '';
-						$matches[7][] = $matches2[1][$ind];
-					}
-				}
-
-				// Try to find all the images!
-				if (!empty($matches))
-				{
-					$image_count_holder = [];
-					foreach ($matches[0] as $key => $image)
-					{
-						$width = -1;
-						$height = -1;
-						$img_count++;
-
-						// Too many images?
-						if (!empty($sig_limits[3]) && $img_count > $sig_limits[3])
-						{
-							// If we've already had this before, we only want to remove the excess.
-							if (isset($image_count_holder[$image]))
-							{
-								$img_offset = -1;
-								$rep_img_count = 0;
-								while ($img_offset !== false)
-								{
-									$img_offset = strpos($sig, $image, $img_offset + 1);
-									$rep_img_count++;
-									if ($rep_img_count > $image_count_holder[$image])
-									{
-										// Only replace the excess.
-										$sig = substr($sig, 0, $img_offset) . str_replace($image, '', substr($sig, $img_offset));
-
-										// Stop looping.
-										$img_offset = false;
-									}
-								}
-							}
-							else
-							{
-								$replaces[$image] = '';
-							}
-
-							continue;
-						}
-
-						// Does it have predefined restraints? Width first.
-						if ($matches[6][$key])
-						{
-							$matches[2][$key] = $matches[6][$key];
-						}
-
-						if ($matches[2][$key] && $sig_limits[5] && $matches[2][$key] > $sig_limits[5])
-						{
-							$width = $sig_limits[5];
-							$matches[4][$key] *= $width / $matches[2][$key];
-						}
-						elseif ($matches[2][$key])
-						{
-							$width = $matches[2][$key];
-						}
-
-						// ... and height.
-						if ($matches[4][$key] && $sig_limits[6] && $matches[4][$key] > $sig_limits[6])
-						{
-							$height = $sig_limits[6];
-							if ($width != -1)
-							{
-								$width *= $height / $matches[4][$key];
-							}
-						}
-						elseif ($matches[4][$key])
-						{
-							$height = $matches[4][$key];
-						}
-
-						// If the dimensions are still not fixed - we need to check the actual image.
-						if (($width == -1 && $sig_limits[5]) || ($height == -1 && $sig_limits[6]))
-						{
-							// We'll mess up with images, who knows.
-							require_once(SUBSDIR . '/Attachments.subs.php');
-
-							$sizes = url_image_size($matches[7][$key]);
-							if (is_array($sizes))
-							{
-								// Too wide?
-								if ($sizes[0] > $sig_limits[5] && $sig_limits[5])
-								{
-									$width = $sig_limits[5];
-									$sizes[1] *= $width / $sizes[0];
-								}
-
-								// Too high?
-								if ($sizes[1] > $sig_limits[6] && $sig_limits[6])
-								{
-									$height = $sig_limits[6];
-									if ($width == -1)
-									{
-										$width = $sizes[0];
-									}
-									$width *= $height / $sizes[1];
-								}
-								elseif ($width != -1)
-								{
-									$height = $sizes[1];
-								}
-							}
-						}
-
-						// Did we come up with some changes? If so, remake the string.
-						if ($width != -1 || $height != -1)
-						{
-							$replaces[$image] = '[img' . ($width != -1 ? ' width=' . round($width) : '') . ($height != -1 ? ' height=' . round($height) : '') . ']' . $matches[7][$key] . '[/img]';
-						}
-
-						// Record that we got one.
-						$image_count_holder[$image] = isset($image_count_holder[$image]) ? $image_count_holder[$image] + 1 : 1;
-					}
-
-					if (!empty($replaces))
-					{
-						$sig = str_replace(array_keys($replaces), array_values($replaces), $sig);
-					}
-				}
-			}
-
-			// Try to fix disabled tags.
-			if (!empty($disabledTags))
-			{
-				$sig = preg_replace('~\[(?:' . implode('|', $disabledTags) . ').+?\]~i', '', $sig);
-				$sig = preg_replace('~\[/(?:' . implode('|', $disabledTags) . ')\]~i', '', $sig);
-			}
-
-			$sig = strtr($sig, ["\n" => '<br />']);
-			call_integration_hook('integrate_apply_signature_settings', [&$sig, $sig_limits, $disabledTags]);
-			if ($sig != $row['signature'])
-			{
-				$changes[$row['id_member']] = $sig;
-			}
-		}
-
-		// Do we need to delete what we have?
-		if (!empty($changes))
-		{
-			foreach ($changes as $id => $sig)
-			{
-				updateSignature($id, $sig);
-			}
-		}
-
-		$applied_sigs += 50;
-		if (!$done)
-		{
-			pauseSignatureApplySettings($applied_sigs);
-		}
-	}
-}
-
 /**
  * Callback for createList() in displaying profile fields
  * Can be used to load standard or custom fields by setting the $standardFields flag
@@ -918,44 +584,4 @@ function scanFileSystemForControllers($iterator, $namespace = '')
 	}
 
 	return $types;
-}
-
-/**
- * Pause the signature applying thing.
- *
- * @param int $applied_sigs
- * @todo Merge with other pause functions?
- *    pausePermsSave(), pauseAttachmentMaintenance(), pauseRepairProcess()
- *
- * @todo Move to subs file
- */
-function pauseSignatureApplySettings($applied_sigs)
-{
-	global $context, $txt, $time_start;
-
-	// Try to get more time...
-	detectServer()->setTimeLimit(600);
-
-	// Have we exhausted all the time we allowed?
-	if ((microtime(true) - $time_start) > 3)
-	{
-		return;
-	}
-
-	$context['continue_get_data'] = '?action=admin;area=featuresettings;sa=sig;apply;step=' . $applied_sigs . ';' . $context['session_var'] . '=' . $context['session_id'];
-	$context['page_title'] = $txt['not_done_title'];
-	$context['continue_post_data'] = '';
-	$context['continue_countdown'] = '2';
-	$context['sub_template'] = 'not_done';
-
-	// Specific stuff to not break this template!
-	$context[$context['admin_menu_name']]['current_subsection'] = 'sig';
-
-	// Get the right percent.
-	$context['continue_percent'] = round(($applied_sigs / $context['max_member']) * 100);
-
-	// Never more than 100%!
-	$context['continue_percent'] = min($context['continue_percent'], 100);
-
-	obExit();
 }
